# Make all images implement IDisposable