… src/ImageSharp.Drawing/project.json
… src/ImageSharp.Formats.Bmp/project.json
… src/ImageSharp.Formats.Gif/project.json
… src/ImageSharp.Formats.Jpeg/project.json
… src/ImageSharp.Formats.Png/project.json
… src/ImageSharp.Processing/Processors/Convolution/EdgeDetection/EdgeDetectorCompassProcessor.cs
… src/ImageSharp.Processing/Processors/Transforms/CompandingResizeProcessor.cs
… src/ImageSharp.Processing/Processors/Transforms/EntropyCropProcessor.cs
… src/ImageSharp.Processing/Processors/Transforms/FlipProcessor.cs
… src/ImageSharp.Processing/Processors/Transforms/ResizeProcessor.cs
… src/ImageSharp.Processing/Processors/Transforms/RotateProcessor.cs
… src/ImageSharp.Processing/Processors/Transforms/SkewProcessor.cs
… src/ImageSharp.Processing/project.json
… src/ImageSharp/Image/IImageBase{TColor}.cs
… src/ImageSharp/Image/ImageBase{TColor}.cs
… src/ImageSharp/Image/Image{TColor}.cs
… src/ImageSharp/Image/PixelAccessor{TColor}.cs
… src/ImageSharp/Image/PixelPool{TColor}.cs
… src/ImageSharp/Quantizers/Quantize.cs
… src/ImageSharp/project.json
… tests/ImageSharp.Benchmarks/Drawing/DrawBeziers.cs
… tests/ImageSharp.Benchmarks/Drawing/DrawLines.cs
… tests/ImageSharp.Benchmarks/Drawing/DrawPolygon.cs
… tests/ImageSharp.Benchmarks/Drawing/FillPolygon.cs
… tests/ImageSharp.Benchmarks/Drawing/FillRectangle.cs
… tests/ImageSharp.Benchmarks/Drawing/FillWithPattern.cs
… tests/ImageSharp.Benchmarks/Image/CopyPixels.cs
… tests/ImageSharp.Benchmarks/Image/DecodeBmp.cs
… tests/ImageSharp.Benchmarks/Image/DecodeFilteredPng.cs
… tests/ImageSharp.Benchmarks/Image/DecodeGif.cs