### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/JimBobSquarePants/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practise as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description

This is a biggun' that's been a long time coming. 

The PR changes anything implementing `IImageBase<TColor>` to now also implement `IDisposable`.

There's a little bit of slight of hand going on within `SetPixels` and `ClonePixels` that allows all this to happen. Nothing too fancy though so it should all be easy to follow. The API is still simple but images should now be wrapped in `using` statements.

```
using (FileStream stream = File.OpenRead("foo.jpg"))
using (FileStream output = File.OpenWrite("bar.jpg"))
using (Image image = new Image(stream))
{
    image.Resize(image.Width / 2, image.Height / 2)
         .Grayscale()
         .Save(output);
}
```

Renting and returning of `TColor[]` instances is managed through a single place. `PixelPool<TColor>`. This public static class should always be the single point where all pixel array pooling is managed. (Note, you'll see an empty catch in `ReturnPixels()` that allows us to handle cases where someone doesn't rent from that pool which should be rare enough to not affect performance.)

The memory differences are astounding. Check out the resize benchmark.

```
BenchmarkDotNet=v0.10.1, OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-6600U CPU 2.60GHz, ProcessorCount=4
Frequency=2742188 Hz, Resolution=364.6723 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1586.0
  DefaultJob : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1586.0
```

### Before

Method |        Mean |    StdErr |    StdDev |      Median | Scaled | Scaled-StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
---------------------------- |------------ |---------- |---------- |------------ |------- |-------------- |--------- |--------- |--------- |---------- |
'System.Drawing Resize' |  76.8127 ms | 1.5556 ms | 9.9606 ms |  71.4049 ms |   1.00 |          0.00 |        - |        - |        - |     512 B |
'ImageSharp Resize' |  74.0430 ms | 0.1690 ms | 0.6095 ms |  73.8233 ms |   0.98 |          0.11 | 329.1667 | 266.6667 | 266.6667 |  20.05 MB |
'ImageSharp Compand Resize' | 126.1365 ms | 0.1987 ms | 0.7435 ms | 125.9578 ms |   1.67 |          0.19 | 333.3333 | 270.8333 | 270.8333 |  20.05 MB |

### After

Method |        Mean |    StdDev | Scaled | Scaled-StdDev | Allocated |
---------------------------- |------------ |---------- |------- |-------------- |---------- |
'System.Drawing Resize' |  72.3372 ms | 1.5962 ms |   1.00 |          0.00 |     512 B |
'ImageSharp Resize' |  70.6951 ms | 0.3417 ms |   0.98 |          0.02 | 186.24 kB |
'ImageSharp Compand Resize' | 120.9624 ms | 0.6795 ms |   1.67 |          0.04 | 186.62 kB |


Could you all please go through this with a fine-tooth comb and ensure I haven't missed anything that could lead to memory leaks. All my tests are passing and I can't find any leaks so far.