@tocsoft 2017-02-01
My initial thought would be that SetPixels should probably copy the passed in TColor[] into is internal memory that means that we never have to deal with pixel data that is owned by someone else when disposing of `ImageBase` which would prevent the need to catch the return in the array pool. it might be a tiny bit slower but it would be a lot safer.

It would also mean that the worry about leaks would be massively reduced as the rule of thumb can then be if you rent it you return it unlike in Convolution2DProcessor.cs where your renting a TColor[] but there's no obvious way in which its collected.

You could also have `PixelPool<TColor>` be able to return a `PixelAccessor<TColor>` would prevent the need to create an array then lock it and then return the rented array back, it would all be encapsulated in the `PixelAccessor<TColor>` which can get its own memory and release it in its dispose.