@JimBobSquarePants 2017-02-01
>you could even have a version of setpixels that is optimized to just swap the backing data from the new pooled pixel accessor and the image that would prevent the copy and mean that the old data would be freed up with the pooled accessor (would have to be careful not to do this when both sets of data are needed to be retained, but if the method was internal and was called something else then it shouldn't be a problem)

@tocsoft  Isn't that exactly what I'm doing?

https://github.com/JimBobSquarePants/ImageSharp/blob/41f96faa7b23fb51463541231f0b0faf5aaa1f1c/src/ImageSharp/Image/ImageBase%7BTColor%7D.cs#L163