@tocsoft 2017-02-01
you are sort of the problem is your doing it in such a way that its hard to track who is currently in charge of what memory... at the moment it looks like Convolution2DProcessor.cs is leaking because your renting but never releasing within the same class/method. Your obviously not leaking there, but it looks like a leak, thus making it harder to maintain.

I'll code up what I mean shortly and show you...seeing code is easier that trying to describe it. 😄