@KLuuKer 2017-02-01
@JimBobSquarePants maybe put a debug only line in that empty catch to throw when someone is returning non-pooled array's, so they will know that they are doing something wrong