@antonfirsov 2017-02-01
@KLuuKer I don't think it's a good idea. `SetPixels()` is a public function, users consuming it from a release binary will not notice they are doing something wrong.

@JimBobSquarePants 
Looking forward for @tocsoft -s gist, maybe his work fill change my mind a bit but currently I'm very concerned, this PR would leave a fundamental problem that must be fixed in the near future: 

*`SetPixels()` is no longer maintainable as a safe API in it's current form! Specially not as a public API.*

I would rather delegate the ownership management of the buffers to a separate `PixelStorageProvider`-like abstract class as commented under #88. It should care about the concern of safely swapping the buffer under an image. It will also enable supporting many custom scenarios in the future.

I would be happy to create a PR for this after #97 is merged!