@JimBobSquarePants 2017-02-01
So it turns out there is no error thrown after all! I could have sworn it used to. That means the catch block is gone.

>SetPixels() is no longer maintainable as a safe API in it's current form! Specially not as a public API.

@antonfirsov Could you expand upon that?