@antonfirsov 2017-02-01
It's a leaky abstraction now:
It's intended to be used used with arrays rented from `PixelPool<TColor>`, but it's not enforced, nor verified when we call `SetPixels()`. We will know we made something wrong only when Dispoese() throws an exception internal to `ArrayPool`. (Worse, if not throwing anything ...)