@antonfirsov 2017-02-01
@JimBobSquarePants it's not well documented but as I understand there [must be an exception thrown](https://github.com/dotnet/corefx/blob/master/src/System.Buffers/src/System/Buffers/DefaultArrayPoolBucket.cs#L91), when you return a non-pooled array.