@JimBobSquarePants 2017-02-01
@antonfirsov I was convinced it was but it isn't and I don't know why

https://github.com/JimBobSquarePants/ImageSharp/blob/94ca0779e3e0c7842742df9bb64c6a405af08a74/tests/ImageSharp.Tests/Image/PixelPoolTests.cs#L48