@benaadams 2017-02-01
It throws if the array length is not a power of 2; as its definitely not from the pool - otherwise the pool can't tell