@tocsoft 2017-02-01
ok had my play see 7940683d37b9259871d527f578b324d3ced9082d for how I propose we change this PR.

This way it means that most people don't even need to think about the PixelPool it just work for them.

Sorry but there ended being a lot more changes than I wanted but it ends up dropping the CopyPixels/SetPixel calls all together.