@tocsoft 2017-02-01
basically my changes moves the responsibility for Renting and Returning the `TColor[]` arrays to the `PixelAccessor<TColor>` & `ImageBase<TColor>` classes thus leaving us with alot less chance of leaking them, plus as they are both IDisposable now you can do away with the `try{}finaly{}`'s as long as you wrap them in `using(..){...}` statements.