@JimBobSquarePants 2017-02-01
@tocsoft I like that, much less leaky. We'd want `internal void SwapPixelsBuffers` to be protected rather than internal so devs can use it though would we not?