@tocsoft 2017-02-01
Not sure about internal vs protected. However my gut says make everything internal/private until a third party dev needs access to it then decide on the best way to make it public then.