@JimBobSquarePants 2017-02-01
@tocsoft Ok then, we've still got plenty of time to review that anyway. Can you post the resize benchmark results under your version please (should be the same) and if everyone's happy then you can update the PR with your changes.