@antonfirsov 2017-02-01
@JimBobSquarePants a non-throwing could be even worse: the user makes a mistake by using [SetPixels() with a non buffered array](https://github.com/JimBobSquarePants/ImageSharp/blob/IDisposable/src/ImageSharp.Processing/Processors/Convolution/ConvolutionProcessor.cs#L47) and he will not notice it at all!

@tocsoft Good job, `SwapPixelsBuffers()` is much better than `SetPixels()` was! This way it's not blinking in red in my eyes anymore! :) 

However I feel buffer management/swapping is not a `PixelAccessor` concern, that's why I want to think about moving this logic to a separate class later.