@tocsoft 2017-02-01
### Before - master 8e56aaaf68177d6e6a14c5dee2a7d4772452895c
``` ini

BenchmarkDotNet=v0.10.1, OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-4770 CPU 3.40GHz, ProcessorCount=8
Frequency=3312645 Hz, Resolution=301.8736 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1586.0
  DefaultJob : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1586.0


```
|                      Method |       Mean |    StdDev | Scaled | Scaled-StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|---------------------------- |----------- |---------- |------- |-------------- |--------- |--------- |--------- |---------- |
|     'System.Drawing Resize' | 59.0057 ms | 2.2540 ms |   1.00 |          0.00 |        - |        - |        - |     512 B |
|         'ImageSharp Resize' | 41.1233 ms | 0.5927 ms |   0.70 |          0.03 | 195.8333 | 195.8333 | 195.8333 |  20.08 MB |
| 'ImageSharp Compand Resize' | 64.2315 ms | 0.6072 ms |   1.09 |          0.04 | 350.0000 | 350.0000 | 350.0000 |  20.08 MB |

### After - 7940683d37b9259871d527f578b324d3ced9082d

``` ini

BenchmarkDotNet=v0.10.1, OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-4770 CPU 3.40GHz, ProcessorCount=8
Frequency=3312645 Hz, Resolution=301.8736 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1586.0
  DefaultJob : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1586.0


```
|                      Method |       Mean |    StdDev | Scaled | Scaled-StdDev | Allocated |
|---------------------------- |----------- |---------- |------- |-------------- |---------- |
|     'System.Drawing Resize' | 57.2351 ms | 0.6135 ms |   1.00 |          0.00 |     512 B |
|         'ImageSharp Resize' | 39.3286 ms | 1.0688 ms |   0.69 |          0.02 | 186.33 kB |
| 'ImageSharp Compand Resize' | 63.1998 ms | 0.9299 ms |   1.10 |          0.02 | 186.36 kB |